virus total flask implementation/requirements.txt
flake8==3.8.1
Flask==1.1.2
Flask-WTF==0.14.3
requests==2.24.0
jsonify==0.5
click==7.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
virustotal-api==1.1.11